pyproject.toml
[project]
name = "trace-invest"
version = "0.1.0"
description = "Systematic trading & investing engine"
requires-python = ">=3.10"

dependencies = [
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "python-dateutil>=2.9.0",
    "pytz>=2024.1",
    "loguru>=0.7.2"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.1.0",
    "ruff>=0.2.1",
    "mypy>=1.8.0"
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
strict = false
